3.0 (#70)
* [vcs_only] - Added 'nppBackup' to .gitignore

* Version bump

* [vcs_only] - update .gitignore

* [vcs_only] - removed .idea folder

* code upgrade

* Add helper code for use with InsPy-Logger

* Complete refactor:
  - Class 'InspyLogger' to be deprecated
  - Add 'Logger' to replace it.
  - Added metaclass 'Loggable' to easily initialize loggers

* - Removed dependency 'domain-suffixes'

* [vcs_only] - Removed outdated and unnecessary documentation

* [vcs_only] - Removed debris

* [vcs_only] - Added 'app.log' to the .gitignore file

* (tag) Version bump 3.0-dev1 -> 3.0-dev2

* Updated requirements.txt

* (add) Added 'inspy-logger-tool' script.

* Added ability to silence InspyLogger on import.

* chore(version): update version to 3.0-dev.3

Update the version number in `version.py` and `pyproject.toml` to 3.0-dev.3.

The updated version reflects the latest development state of the project.

* refactor(helpers): refactor Loggable class and add base_classes module

The Loggable class in the helpers module has been refactored to improve its functionality and readability. The class now inherits from a new LoggableDescriptor class defined in the base_classes module. This descriptor allows for easy access to a logger specific to a class method. Additionally, a new _get_parent_logging_device() function has been added to determine the parent logging device by inspecting the caller's log_device or parent_log_device attribute.

These changes improve the logging capabilities of the Loggable class and enhance consistency in logging behavior and hierarchy.

* feat(examples): add loggable_class and silent_import examples

Add two new example files, loggable_class.py and silent_import.py, to the
examples/v3.0 directory. These files demonstrate how to use the Loggable class
and the silent import feature of the logger.

Both files contain code that sets up a logger instance and logs debug and info
messages. The loggable_class.py file also includes a LoggedClass class that
extends the Loggable class and demonstrates how to use a child logger within a
class.

This addition enhances the examples provided in the project and helps users
understand how to use the logger effectively.

Issue: 3.0-DEV.3

* refactor(docs): add Sphinx documentation files

Added the following files to the docs folder:
- Makefile
- make.bat
- index.rst
- conf.py

These files are necessary for generating the Sphinx documentation for the Inspy-Logger project.

* :sparkles: feat(inspy_logger): add detailed method docstrings and remove unnecessary lines

* :art: style: use f-string and remove extra newline in example files

* :recycle: refactor(inspy_logger): Use `_log` method for logging calls

:sparkles: feat: Add `.readthedocs.yaml` for documentation configuration

:lipstick: style(main.py): Simplify `main` function logic and fix string formatting

* .

* .

* :arrow_up: :bookmark: Bump version number to 3.0.0-dev.4 and update dependencies

* :arrow_up: :bookmark: Bump version number to 3.0.0-dev.4 and update dependencies

* :bookmark: feat: Update inspy-logger version to 3.0, add error handling in main.py, add new dev dependency in pyproject.toml

* :bookmark: feat(version): Update version from 3.0 to 3.0.1 in pyproject.toml, version.py and setup.py

* Update python-package.yml

* :arrow_up: chore: update python and os versions in GitHub workflow

* :gear: chore: Update OS version in GitHub workflow

* :arrow_up: chore: update macos version in python-package.yml workflow

* :gear: chore(workflows): Update OS matrix in python-package workflow

* 'Refactored by Sourcery' (#69)

Co-authored-by: Sourcery AI <>

---------

Co-authored-by: sourcery-ai[bot] <58596630+sourcery-ai[bot]@users.noreply.github.com>