feat(debug): Add system information gathering helpers
This commit adds several helpers for gathering system information. This
includes information about the operating system, processor, and
execution environment. The information can be formatted in several
formats including markdown, json, and text. It can also be copied
directly to the clipboard.

This commit also includes updates to the version checking and update
checking functionality, as well as the addition of a new `ByteConverter`
class for converting between various units of digital storage.

In addition to these, several minor updates and additions are included
in this commit, such as updates to the package version and additional
dependencies.