refactor(inspy-logger): streamline and tidy up logger code
In this commit, a variety of changes were made to streamline and tidy up the logger code. Unnecessary code blocks were removed, new functions were added, and overall structure was improved. This has resulted in cleaner, more readable code which should be easier to maintain and debug in the future. The version was also updated to 3.0.2. The benefits of these changes include improved readability and maintainability, as well as a more efficient logger.